parlink/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
click==7.0
configparser==3.7.4
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
flask==1.0.3
future==0.17.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
importlib-metadata==0.18
ipykernel==5.1.1
ipython-genutils==0.2.0
ipython==7.5.0 ; python_version >= '3.3'
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.14.0
jieba==0.39
jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyter==1.0.0
line-bot-sdk==1.12.1
lxml==4.3.4
markupsafe==1.1.1
mistune==0.8.4
more-itertools==7.1.0 ; python_version > '2.7'
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.4.3
notebook==5.7.8
oauthlib==3.0.1
packaging==19.0
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0 ; os_name != 'nt'
py==1.8.0
pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
pysocks==1.7.0
pytest==4.6.3
python-dateutil==2.8.0
pyzmq==18.0.1
qtconsole==4.5.1
requests-oauthlib==1.2.0
requests==2.22.0
send2trash==1.5.0
six==1.12.0
snownlp==0.12.3
terminado==0.8.2
testpath==0.4.2
textblob==0.15.3
tornado==6.0.3
traitlets==4.3.2
tweepy==3.7.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
werkzeug==0.15.4
widgetsnbextension==3.4.2
zipp==0.5.1
pandas==0.24.2